+++ src/main/java/com/rkmd/toki_no_nagare/configuration/AppConfig.java
@@ -3,8 +3,10 @@ package com.rkmd.toki_no_nagare.configuration;
 import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @Configuration
+@EnableScheduling
 public class AppConfig {
 
   @Bean